add basic dataLayer properties
